// src/src/core/constants/enums.ts
export enum Reference {
  Unused = 0,
  Used = 1,
}
// ...
export enum Status {
  Active = 1,
  Inactive = 0,
  Probation = 2,
  Suspended = 3,
}
// ...
export enum BookingStatus {
  Pending = 0,
  Booked = 1,
}

export enum TestStatus {
  Scheduled = 'scheduled',
  ReScheduled = 'rescheduled',
  Failed = 'failed',
  Passed = 'passed',
  Missed = 'missed',
}

export enum PaymentStatus {
  Pending = 'pending',
  Completed = 'completed',
  Failed = 'failed',
  Canceled = 'canceled',
  Used = 'used',
}